refactor: update documentation components and improve code organization
- Replaced instances of CodeMdx with CodeHighlight in various documentation files for consistency.
- Updated import paths for documentation navigation components to reflect new structure.
- Removed deprecated documentation components and streamlined the overall codebase.
- Adjusted package.json and bun.lock to remove unused dependencies and ensure clarity in project structure.
