+++ content/components/input-field.mdx
@@ -15,7 +15,7 @@ group: form
 
 With classic shadcn/ui, creating a form field requires a lot of repetitive code:
 
-<CodeMdx code={`// With classic shadcn/ui - verbose
+<CodeHighlight code={`// With classic shadcn/ui - verbose
 <FormField
   control={form.control}
   name="email"
@@ -35,7 +35,7 @@ With classic shadcn/ui, creating a form field requires a lot of repetitive code:
 
 With `InputField`, the same result in a single line:
 
-<CodeMdx code={`// With InputField - concise
+<CodeHighlight code={`// With InputField - concise
 <InputField 
   control={form.control} 
   name="email" 
@@ -56,7 +56,7 @@ With `InputField`, the same result in a single line:
 
 The component is specifically designed for react-hook-form:
 
-<CodeMdx code={`const form = useForm({
+<CodeHighlight code={`const form = useForm({
   defaultValues: { name: '', age: 0 },
   resolver: zodResolver(schema),
 });
