+++ content/docs/configuration.mdx
@@ -12,7 +12,7 @@ shuip components are designed to be **easily customizable**. Each component acce
 
 The simplest way to customize a component:
 
-<CodeMdx code={`import { SubmitButton } from '@/components/ui/shuip/submit-button'
+<CodeHighlight code={`import { SubmitButton } from '@/components/ui/shuip/submit-button'
 // ...
 <SubmitButton 
   	label="Send"
@@ -25,7 +25,7 @@ The simplest way to customize a component:
 
 Since the components are copied directly into your project, you can freely modify them:
 
-<CodeMdx code={`// src/components/ui/shuip/submit-button.tsx
+<CodeHighlight code={`// src/components/ui/shuip/submit-button.tsx
 
 export function SubmitButton({
     onClick,
