+++ content/docs/installation.mdx
@@ -32,7 +32,7 @@ You can also copy and paste the code directly from each component’s documentat
 
 After installation, shuip components will be added to your project like this:
 
-<CodeMdx code={`src/
+<CodeHighlight code={`src/
 ├── components/
 │   ├── block/
 │   │   └── shuip/                 # 🆕 shuip blocks
@@ -51,7 +51,7 @@ After installation, shuip components will be added to your project like this:
 
 If you're using TypeScript, make sure your `tsconfig.json` includes the appropriate path aliases:
 
-<CodeMdx code={`{
+<CodeHighlight code={`{
     "compilerOptions": {
       "baseUrl": ".",
       "paths": {
