+++ package.json
@@ -50,7 +50,6 @@
     "@radix-ui/react-toast": "^1.2.14",
     "@radix-ui/react-tooltip": "^1.2.7",
     "@tailwindcss/postcss": "^4.1.7",
-    "@tanstack/react-query": "^5.77.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "contentlayer2": "0.5.8",
@@ -65,7 +64,6 @@
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.56.4",
     "remark": "^15.0.1",
-    "remark-gfm": "^4.0.1",
     "shadcn": "^2.5.0",
     "tailwind-merge": "^3.3.0",
     "tailwindcss-animate": "^1.0.7",
