+++ src/components/shared/code-preview.tsx
@@ -11,7 +11,7 @@ export interface CodePreviewProps {
   language?: string;
 }
 
-export function CodePreview({ code, filename }: CodePreviewProps) {
+export function CodePreview({ code, filename, language = 'tsx' }: CodePreviewProps) {
   const codeString = code
     ? code
     : React.useMemo(() => {
@@ -25,9 +25,15 @@ export function CodePreview({ code, filename }: CodePreviewProps) {
       }, [filename]);
 
   return (
-    <div className='w-full max-w-[850px] border rounded-lg [&_pre]:my-0 [&_pre]:max-h-[350px] [&_pre]:overflow-auto'>
+    <div className='w-full border rounded-lg [&_pre]:my-0 [&_pre]:max-h-[350px] [&_pre]:overflow-auto'>
       {codeString ? (
-        <CodeHighlight code={codeString} />
+        <div>
+          <div className='flex items-center justify-between p-2 border-b'>
+            {/* TODO V0 */}
+            <ButtonCopy value={codeString} />
+          </div>
+          <CodeHighlight code={codeString} language={language} />
+        </div>
       ) : (
         <p className='text-sm text-muted-foreground'>Component not found in registry.</p>
       )}
@@ -35,31 +41,25 @@ export function CodePreview({ code, filename }: CodePreviewProps) {
   );
 }
 
-interface CodeHighlightProps extends React.HTMLAttributes<HTMLDivElement> {
+export interface CodeHighlightProps {
   code: string;
   language?: string;
 }
 
-function CodeHighlight({ code, language = 'tsx' }: CodeHighlightProps) {
+export function CodeHighlight({ code, language = 'tsx' }: CodeHighlightProps) {
   return (
-    <div>
-      <div className='flex items-center justify-between p-2 border-b'>
-        {/* TODO V0 */}
-        <ButtonCopy value={code} />
-      </div>
-      <Highlight theme={themes.vsDark} code={code} language={language}>
-        {({ style, tokens, getLineProps, getTokenProps }) => (
-          <pre style={style} className='p-4'>
-            {tokens.map((line, i) => (
-              <div key={i} {...getLineProps({ line })}>
-                {line.map((token, key) => (
-                  <span key={key} {...getTokenProps({ token })} />
-                ))}
-              </div>
-            ))}
-          </pre>
-        )}
-      </Highlight>
-    </div>
+    <Highlight theme={themes.vsDark} code={code} language={language}>
+      {({ style, tokens, getLineProps, getTokenProps }) => (
+        <pre style={style} className='p-4'>
+          {tokens.map((line, i) => (
+            <div key={i} {...getLineProps({ line })}>
+              {line.map((token, key) => (
+                <span key={key} {...getTokenProps({ token })} />
+              ))}
+            </div>
+          ))}
+        </pre>
+      )}
+    </Highlight>
   );
 }
